% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addArg.R
\name{addArg}
\alias{addArg}
\title{addArg}
\usage{
addArg(parser, cat = c("General", "Plot"))
}
\arguments{
\item{method2order}{method to order colors (\code{"hsv"} or \code{"cluster"})}
}
\value{
None
}
\description{
Description: An wrapper function that add options for multiple categories
}
\examples{
parser <- addArg(parser,"General")
}
\author{
Rui Chen (crotoc@gmail.com)
}
