% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addArg.R
\name{addArgEnhancer}
\alias{addArgEnhancer}
\title{addArgEnhancer}
\usage{
addArgEnhancer(parser)
}
\arguments{
\item{method2order}{method to order colors (\code{"hsv"} or \code{"cluster"})}
}
\value{
None
}
\description{
Description: Add Enhancer related options
}
\examples{
addArgEnhancer(parser)

}
\author{
Rui Chen (crotoc@gmail.com)
}
