% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addArg.R
\name{addArgGeneral}
\alias{addArgGeneral}
\title{addArgGeneral}
\usage{
addArgGeneral(parser)
}
\arguments{
\item{method2order}{method to order colors (\code{"hsv"} or \code{"cluster"})}
}
\value{
None
}
\description{
Description: Add options for general purpose through argparse package
}
\examples{
parser <- ArgumentParser()
parser <- addArgGeneral(parser)

}
\author{
Rui Chen (crotoc@gmail.com)
}
