% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addArg.R
\name{addArgGwas}
\alias{addArgGwas}
\title{addArgGwas}
\usage{
addArgGwas(parser)
}
\arguments{
\item{method2order}{method to order colors (\code{"hsv"} or \code{"cluster"})}
}
\value{
None
}
\description{
Description: Add options related to GWAS file
}
\examples{
parser <- addArgGwas(parser)

}
\author{
Rui Chen (crotoc@gmail.com)
}
