% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addArg.R
\name{addArgHeatmap}
\alias{addArgHeatmap}
\title{addArgHeatmap}
\usage{
addArgHeatmap(parser)
}
\arguments{
\item{method2order}{method to order colors (\code{"hsv"} or \code{"cluster"})}
}
\value{
None
}
\description{
Description: add heatmap options
}
\examples{
addArgHeatmap(parser)

}
\author{
Rui Chen (crotoc@gmail.com)
}
