% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addArg.R
\name{addArgHighc}
\alias{addArgHighc}
\title{addArgHighc}
\usage{
addArgHighc(parser)
}
\arguments{
\item{method2order}{method to order colors (\code{"hsv"} or \code{"cluster"})}
}
\value{
None
}
\description{
Description: add HighC related options
}
\examples{
addArgHighc(parser)

}
\author{
Rui Chen (crotoc@gmail.com)
}
