% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addArg.R
\name{addArgIrigs}
\alias{addArgIrigs}
\title{addArgIrigs}
\usage{
addArgIrigs(parser)
}
\arguments{
\item{method2order}{method to order colors (\code{"hsv"} or \code{"cluster"})}
}
\value{
None
}
\description{
Description: add options related to iRIGS
}
\examples{
parser <- addArgIrigs(parser)

}
\author{
Rui Chen (crotoc@gmail.com)
}
