% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addArg.R
\name{addArgPlot}
\alias{addArgPlot}
\title{addArgPlot}
\usage{
addArgPlot(parser)
}
\arguments{
\item{method2order}{method to order colors (\code{"hsv"} or \code{"cluster"})}
}
\value{
None
}
\description{
Description: Add options related to ggplot2 ploting
}
\examples{
parser <- addArgPlot(parser)

}
\author{
Rui Chen (crotoc@gmail.com)
}
