% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addArg.R
\name{addArgRmd}
\alias{addArgRmd}
\title{addArgRmd}
\usage{
addArgRmd(parser)
}
\arguments{
\item{method2order}{method to order colors (\code{"hsv"} or \code{"cluster"})}
}
\value{
None
}
\description{
Description: Add options related to runRmd.R script
}
\examples{
parser <- addArgRmd(parser)

}
\author{
Rui Chen (crotoc@gmail.com)
}
