% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addArg.R
\name{addArgScenic}
\alias{addArgScenic}
\title{addArgScenic}
\usage{
addArgScenic(parser)
}
\arguments{
\item{method2order}{method to order colors (\code{"hsv"} or \code{"cluster"})}
}
\value{
None
}
\description{
Description: Add options related to run SCENIC
}
\examples{
parser <- addArgScenic(parser)

}
\author{
Rui Chen (crotoc@gmail.com)
}
