% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addArg.R
\name{addArgSinglecell}
\alias{addArgSinglecell}
\title{addArgSinglecell}
\usage{
addArgSinglecell(parser)
}
\arguments{
\item{method2order}{method to order colors (\code{"hsv"} or \code{"cluster"})}
}
\value{
None
}
\description{
Description: Add options related to single cell analysis using Seurat
}
\examples{
parser <- addArgSinglecell(parser)

}
\author{
Rui Chen (crotoc@gmail.com)
}
