% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addArg.R
\name{addArgSnp}
\alias{addArgSnp}
\title{addArgSnp}
\usage{
addArgSnp(parser)
}
\arguments{
\item{method2order}{method to order colors (\code{"hsv"} or \code{"cluster"})}
}
\value{
None
}
\description{
Description: Add options reltead to snp file parsing
}
\examples{
parser <- addArgSnp(parser)

}
\author{
Rui Chen (crotoc@gmail.com)
}
