% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addArg.R
\name{addArgSumstats}
\alias{addArgSumstats}
\title{addArgSumstas}
\usage{
addArgSumstats(parser)
}
\arguments{
\item{method2order}{method to order colors (\code{"hsv"} or \code{"cluster"})}
}
\value{
None
}
\description{
Description: Add options related to summary statistics of GWAS
}
\examples{
parser <- addArgSumstas(parser)

}
\author{
Rui Chen (crotoc@gmail.com)
}
