% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGeom.R
\name{addGeom}
\alias{addGeom}
\title{addGeom}
\usage{
addGeom(x_plot, type, opt, ...)
}
\arguments{
\item{no}{param need.}
}
\value{
None
}
\description{
add geom on the existing plot
}
\examples{
addGeom(x_plot,type,opt)


}
\author{
Rui Chen (crotoc@gmail.com)
}
