% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendWithOverlap.R
\name{appendWithOverlap.xts}
\alias{appendWithOverlap.xts}
\title{appendWithOverlap.xts}
\usage{
\method{appendWithOverlap}{xts}(x, y)
}
\arguments{
\item{x}{}

\item{y}{}
}
\description{
appendWithOverlap.xts
}
