% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apsimng.R
\name{apsimng_cultivars}
\alias{apsimng_cultivars}
\title{Get releae cultivars from apsimx}
\usage{
apsimng_cultivars(crop)
}
\arguments{
\item{crop}{The crop name}
}
\value{
A data frame for all cultivar name and parameter values
}
\description{
Get releae cultivars from apsimx
}
