% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_morlet.R
\name{calc_phase_angle}
\alias{calc_phase_angle}
\title{calculate phase from mat_wt}
\usage{
calc_phase_angle(mat_wt)
}
\arguments{
\item{mat_wt}{matrix containing the wavelet transform}
}
\value{
the time series of phase angles
}
\description{
calculate phase from mat_wt
}
