% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{capitalize}
\alias{capitalize}
\title{Capitalize a Character Vector}
\usage{
capitalize(string, all = F)
}
\arguments{
\item{string}{a character or character vector}

\item{all}{logical, defauls to \code{FALSE}. If \code{TRUE}, will capitalize all words in each entry. Otherwise, only capitalizes first word.}
}
\value{
A character or character vector with either the first or all words capitalized.
}
\description{
Capitalize a Character Vector
}
\examples{
capitalize("hello world") # "Hello world"
capitalize(c("hello world", "hello mom")) # c("Hello world", "Hello mom")
capitalize(c("hello world"), all = T) # "Hello World"
capitalize(c("hello world", "hello mom"), all = T) # c("Hello World", "Hello Mom")
}
