% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physics.R
\name{CD_sphere}
\alias{CD_sphere}
\title{Drag coefficient of a spherical object}
\usage{
CD_sphere(d_s, v_f, rho_f, mu_f, retall = FALSE)
}
\arguments{
\item{d_s}{diameter of a sphere - m}

\item{v_f}{velocity - m/s}

\item{rho_f}{density of fluid - kg/m3}

\item{mu_f}{fluid viscosity - Pa-s}

\item{retall}{description}
}
\description{
Drag coefficient of a spherical object
}
