% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{chunk}
\alias{chunk}
\title{Generate an vector with size and length}
\usage{
chunk(size, length)
}
\arguments{
\item{size}{size of each chunk}

\item{length}{length of total vector}
}
\value{
A vector
}
\description{
Generate an vector with size and length
}
\examples{
chunk(10, 50)
}
