% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myutils.R
\name{col32}
\alias{col32}
\title{Add alpha level to color}
\usage{
col32(cols, alphas)
}
\arguments{
\item{cols}{vector of color name or hex (e.g. red, #123456)}

\item{alphas}{vector of alpha levels (from 0 to 255) [optional]}
}
\value{
vector of color codes
}
\description{
Add alpha level to color
}
