% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{collapseWS}
\alias{collapseWS}
\title{Collapse all white space blocks into a single space}
\usage{
collapseWS(str)
}
\arguments{
\item{str}{a character or character vector}
}
\value{
A character or character vector with only single spaces
}
\description{
Collapse all white space blocks into a single space
}
\note{
This has a recursive implementation, so should not be used to remove spaces of extreme length
}
\examples{
collapseWS("hello    world") # "hello world"
}
