% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demBlue}
\alias{demBlue}
\alias{demBlueAlt}
\alias{demBlueLight}
\alias{demBlueWhite}
\alias{gopRed}
\alias{gopRedAlt}
\alias{gopRedLight}
\alias{gopRedWhite}
\alias{indGreen}
\alias{indGreenAlt}
\alias{indGreenLight}
\alias{indGreenWhite}
\alias{color-palette}
\title{Color Palette}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
demBlue

demBlueAlt

demBlueLight

demBlueWhite

gopRed

gopRedAlt

gopRedLight

gopRedWhite

indGreen

indGreenAlt

indGreenLight

indGreenWhite
}
\description{
Election graphics color palette including primary and alternate colors for democrats, republicants, and independents. Use \code{demBlueXXX} for democrats, \code{gopRedXXX} for republicans, and \code{indGreenXXX} for others. Primary colors are \code{demBlue}, \code{gopRed}, and \code{indGreen}. Darker alternate colors are given by the suffix \code{Alt}. Lighter accent colors are given by the suffix \code{Light}. Off-whites are given by the suffix \code{White}.
}
\keyword{datasets}
