% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_change_rast_timeseries.R
\name{compute_change_rast_timeseries}
\alias{compute_change_rast_timeseries}
\title{create list of MODIS FC download links}
\usage{
compute_change_rast_timeseries(x)
}
\arguments{
\item{x}{character vector of file location, or terra SpatRaster ordered in a time series}
}
\value{
a single terra raster with each pixel the value of the slope of the lm
}
\description{
author: Ned Ryan-Schofield
date: 5/02/2026
Function for computing slope of linear model for each pixel in a time series raster stack. Output is a single raster with each pixel the value of the slope of the lm
}
