% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_tgi_top_n.R
\name{compute_tgi_top_n}
\alias{compute_tgi_top_n}
\title{Compute dataset TGI}
\usage{
compute_tgi_top_n(x, target, category, tgi_value, top_number = 10)
}
\arguments{
\item{x}{数据集}

\item{target}{目标对象}

\item{category}{分组类型}

\item{tgi_value}{数值}

\item{top_number}{最大排名}
}
\description{
Compute dataset TGI
}
