% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_POSIXct_2_Date.R
\name{convertPOSIXct2Date}
\alias{convertPOSIXct2Date}
\title{convertPOSIXct2Date}
\usage{
convertPOSIXct2Date(x)
}
\arguments{
\item{x}{A xts object}
}
\value{
A converted xts object
}
\description{
convert the index of xts object from POSIXct to Date
}
