% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myUtils.R
\name{cpm}
\alias{cpm}
\title{Counts per million normalization}
\usage{
cpm(data)
}
\arguments{
\item{data}{The count matrix to be normalized}
}
\description{
This function converts an unnormalized countmatrix of genes x samples into counts-per-million normalization
}
\examples{
cpm()
}
\keyword{normalization}
