% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinformatics.R
\name{crop_chromosome}
\alias{crop_chromosome}
\title{Lists of chromosome for crops}
\usage{
crop_chromosome(crops, type = c("list", "data.frame"))
}
\arguments{
\item{crops}{a vector of crops (i.e. wheat, barley, canola)}

\item{type}{return as list or data.frame}
}
\value{
a list or data.frame for predefined chromosome
}
\description{
Lists of chromosome for crops
}
\examples{
crop_chromosome(c("wheat", "barley", "canola"), type = "data.frame")
crop_chromosome(c("wheat", "barley", "canola"))
}
