% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmm_to_dt.R
\name{ctmm_to_dt}
\alias{ctmm_to_dt}
\title{ctmm telemetry to data.table}
\usage{
ctmm_to_dt(telemetry)
}
\arguments{
\item{telemetry}{a ctmm telemetry object}

\item{crs}{projection to use for as.telemetry. e.g. "EPSG:7852" for UTM 52 - tiwis}

\item{x}{numerical variable - data.table coordinate longitude column}

\item{y}{numerical variable - data.table coordinate latitude column}

\item{ID}{character variable - data.table animal ID column}
}
\value{
a list of data.table objects
}
\description{
author: Ned Ryan-Schofield
date: 03/03/2026
Helper function for coercing a ctmm telemetry object back to a list of data.tables, after outlier removal.
}
