% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fip.R
\name{DarcyWeisbach}
\alias{DarcyWeisbach}
\title{Darcy-Weisbach equation}
\usage{
DarcyWeisbach(fD, density, velocity, D)
}
\arguments{
\item{fD}{Darcy friction factor}

\item{density}{density}

\item{velocity}{velocity}

\item{D}{pipe diameter}
}
\value{
dP per unit length
}
\description{
Darcy-Weisbach equation
}
