% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dates.R
\name{date2nbDaysSinceEpid}
\alias{date2nbDaysSinceEpid}
\title{Calculate number of days since start of epidemics}
\usage{
date2nbDaysSinceEpid(date_0)
}
\arguments{
\item{date_0}{a Date}
}
\value{
the number of days
between date in date_0
and start of the corresponding epidemic season
(1st Feb if date is in Feb-Jul,
1st Aug if date is in Aug-Jan)
}
\description{
Calculate number of days since start of epidemics
}
