% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_with_timestamp.R
\name{dir_with_timestamp}
\alias{dir_with_timestamp}
\title{dir_with_timestamp}
\usage{
dir_with_timestamp(
  path = ".",
  pattern = "",
  full.names = FALSE,
  name_timestamp_seperator = "_",
  timestamp_format = "\%Y-\%m-\%d_\%H.\%M.\%S",
  remove_timestamp = FALSE,
  only_most_recent = FALSE,
  ...
)
}
\arguments{
\item{path}{A string with the path to the folder which is to be listed.}

\item{pattern}{A string with a regex pattern with which to filter.}

\item{full.names}{Boolean, if true return full paths.}

\item{name_timestamp_seperator}{A string with the name-timestamp seperator.}

\item{timestamp_format}{A string with the timestamp format, as input for
\code{\link[base:as.POSIXlt]{base::as.POSIXct()}}.}

\item{remove_timestamp}{Boolean, if true remove timestamp.}

\item{only_most_recent}{Boolean, if true remove folders with identical name
but older timestamps.}

\item{...}{Parameters that are passed on to \code{\link[base:list.files]{base::dir()}}.}
}
\value{
A vector of strings with folder paths.
}
\description{
dir_with_timestamp is a wrapper around \code{\link[base:list.files]{base::dir()}}, with some
additional capabilities like filtering out folders with identical names but
old timestamps.
}
