% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{dk.kml3d}
\alias{dk.kml3d}
\title{New kml3d}
\usage{
dk.kml3d(
  object,
  nbClusters = 2:6,
  nbRedrawing = 20,
  toPlot = "none",
  parAlgo = parKml3d()
)
}
\arguments{
\item{object}{C}

\item{nbClusters}{h}

\item{nbRedrawing}{r}

\item{toPlot}{i}

\item{parAlgo}{s}
}
\value{
t
}
\description{
Modified to turn off saving.  I am quite sure it needs kml3d libraried.
}
\examples{
# No thanks.
}
