% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doWithProxy.R
\name{doWithProxy}
\alias{doWithProxy}
\title{doWithProxy}
\usage{
doWithProxy(cmd, proxy_url = "127.0.0.1", port = 1080)
}
\arguments{
\item{cmd}{r commands to be execute with proxy}

\item{port}{}
}
\description{
doWithProxy
}
\examples{
doWithProxy(GET("http://www.google.com"))
}
