% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{doy2date}
\alias{doy2date}
\title{Convert day of year (doy) to date in month and day format}
\usage{
doy2date(doy, format = "\%d-\%b")
}
\arguments{
\item{doy}{a vertor of day of year}

\item{format}{the date format}
}
\value{
Date in month and day format in character format
}
\description{
Convert day of year (doy) to date in month and day format
}
\details{
No leap year is considered
}
\examples{
doy2date(seq(1, 10))
}
