% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_has_no_empty_points.R
\name{dt_has_no_empty_points}
\alias{dt_has_no_empty_points}
\title{check data.table has no empty points}
\usage{
dt_has_no_empty_points(dt, x_col = "long", y_col = "lat")
}
\arguments{
\item{dt}{a list of data.tables of GPS data}

\item{x_col}{name of the x coordinate vector, should be numeric}

\item{y_col}{name of the y coordinate vector, should be numeric}
}
\value{
a list of TRUE or FALSE values
}
\description{
author: Ned Ryan-Schofield
date: 27/02/2026
Helper function for checking whether a list of data.table objects (that are gps data derived using spot_to_DT function) have any empty lat/long positions
example use: lapply(dt, dt_has_no_empty_points)
}
