% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_is_time_ordered_no_duplicates.R
\name{dt_is_time_ordered_no_duplicates}
\alias{dt_is_time_ordered_no_duplicates}
\title{check data.table is time ordered without duplicates}
\usage{
dt_is_time_ordered_no_duplicates(dt, time_col = "timestamp", na_ok = FALSE)
}
\arguments{
\item{dt}{a list of data.tables of GPS data}

\item{time_col}{name of the timestamp vector, of class POSIXct}

\item{na_ok}{an NA check. If true, NA's accepted.}
}
\value{
a list of TRUE or FALSE values
}
\description{
author: Ned Ryan-Schofield
date: 27/02/2026
Helper function for checking whether a list of data.table objects (that are gps data derived using spot_to_DT function) are time ordered, without duplicate timestamps.
example use: lapply(dt, dt_is_time_ordered)
}
