% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_to_amt_track_xyt.R
\name{dt_to_amt_track}
\alias{dt_to_amt_track}
\title{convert data.table to amt track.xyt}
\usage{
dt_to_amt_track(
  dt,
  crs_out,
  lon = "long",
  lat = "lat",
  time = "timestamp",
  id = "ID",
  crs_in = 4326
)
}
\arguments{
\item{crs_out}{crs code. accept 3577, "EPSG:3577", proj string, or WKT}

\item{lon}{character denoting numeric longitude column}

\item{lat}{character denoting numeric latitude column}

\item{time}{character denoting posixct time column}

\item{id}{character denoting ID column}

\item{crs_in}{crs code. accept 3577, "EPSG:3577", proj string, or WKT}

\item{x}{list of data.table objects or single data.table}
}
\value{
a list of track.xyt objects
}
\description{
author: Ned Ryan-Schofield
date: 19/02/2026
Helper function for converting a list of data.table objects to amt_xyt data structure for step selection make sure that you specify the crs as the same as the covariate raster stack
example usage:
x <- lapply(gps, dt_to_amt_track, crs_out = tiwi.crs)
x.cov <- amt::extract_covariates(x, fire)
}
