% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_to_ctmm.R
\name{dt_to_ctmm}
\alias{dt_to_ctmm}
\title{data.table to ctmm telemetry object}
\usage{
dt_to_ctmm(dt, crs, x = "long", y = "lat", ID = "ID")
}
\arguments{
\item{dt}{a list of data.tables of GPS data}

\item{crs}{projection to use for as.telemetry. e.g. "EPSG:7852" for UTM 52 - tiwis}

\item{x}{numerical variable - data.table coordinate longitude column}

\item{y}{numerical variable - data.table coordinate latitude column}

\item{ID}{character variable - data.table animal ID column}
}
\value{
a ctmm telemetry object
}
\description{
author: Ned Ryan-Schofield
date: 28/02/2026
Helper function for coercing a list of data.table objects (that are gps data derived using spot_to_DT function) to ctmm telemetry objects, for easy removal of outliers
}
