% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_to_movebank_tab.R
\name{dt_to_movebank_tab}
\alias{dt_to_movebank_tab}
\title{data.table to movebank textfile}
\usage{
dt_to_movebank_tab(
  dt,
  out.path,
  t = "timestamp",
  y = "lat",
  x = "long",
  ID = "ID",
  tz_in = "Australia/Darwin",
  round_time = TRUE
)
}
\arguments{
\item{dt}{a list of data.tables that are gps data, with same column names}

\item{y}{numerical variable - data.table coordinate latitude column}

\item{x}{numerical variable - data.table coordinate longitude column}

\item{ID}{character variable - data.table animal ID column}

\item{tz_in}{character vector of timezone that timestamp (which is a POSIXct) is in. Default is "Australia/Darwin"}

\item{outpath}{character variable of file path name. i.e. "./outputs/table/tiwi_cats_movebank_up_03032026.txt"}

\item{timestamp}{POSIXct variable - data.table timestamp column}
}
\value{
writes a textfile to disc. Timestamp is in UTC +0
}
\description{
author: Ned Ryan-Schofield
date: 23/04/2026
Helper function for coercing a list of data.table objects of gps data that have been checked for errors and coerced back to data.table to a single .txt file for upload to movebank. Still a bit of a proccess to get this data into movebank though.
This function converts from the local timezone to UTC +0, which is easier to work with in movebank.
}
