% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_CEStree.R
\name{explore_CEStree}
\alias{explore_CEStree}
\title{Explore the CES tree of a REMIND run}
\usage{
explore_CEStree(gdx_filepaths)
}
\arguments{
\item{gdx_filepaths}{A vector of strings to REMIND gdx files}
}
\description{
explore_CEStree opens a shiny app in which the ces parameters and values may
be explored.
}
