% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_morlet.R
\name{extract_scales}
\alias{extract_scales}
\title{Extract components between 2 scales}
\usage{
extract_scales(mat_wt, range_scale, fill = 0)
}
\arguments{
\item{mat_wt}{matrix containing the wavelet transform}

\item{range_scale}{vector of length 2 containing scales to extract}

\item{fill}{variable to fill matrix outside of scale range}
}
\value{
wt matrix with scales in range_scale and zero otherwise
}
\description{
Extract components between 2 scales
}
