% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facToNum.R
\name{facToNum}
\alias{facToNum}
\title{Coerse Factor to Numeric}
\usage{
facToNum(x)
}
\arguments{
\item{x}{a factor}
}
\value{
The factor's names as numerics.
}
\description{
Coerse Factor to Numeric
}
\note{
This is simply a shorthand for a standard converstion to numeric.
}
