% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physics.R
\name{FD}
\alias{FD}
\title{Drag force}
\usage{
FD(CD, rho_f, v, A_ref)
}
\arguments{
\item{CD}{drag coefficient}

\item{rho_f}{fluid density   - kg/m3}

\item{v}{(slip) velocity - m/s}

\item{A_ref}{reference area (projected area) - m2}
}
\value{
drag force - N
}
\description{
Drag force
}
