% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fip.R
\name{fD_laminar}
\alias{fD_laminar}
\title{Darcy friction factor in laminar flow}
\usage{
fD_laminar(Re)
}
\arguments{
\item{Re}{Reynolds number}
}
\value{
Darcy friction factor
}
\description{
Darcy friction factor in laminar flow
}
