% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_morlet.R
\name{find_zeros_lin}
\alias{find_zeros_lin}
\title{plot real part of the wavelet decomposition}
\usage{
find_zeros_lin(x, f_x, bool_inc = T, bool_dec = T)
}
\arguments{
\item{x}{vector containing values of x}

\item{f_x}{vector containing values of functions}

\item{bool_inc}{boolean indicating whether
returning zeros with positive slopes}

\item{bool_dec}{boolean indicating whether
returning zeros with negative slopes}
}
\value{
the zeros of the function
}
\description{
plot real part of the wavelet decomposition
}
