% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{findLastWord}
\alias{findLastWord}
\title{Find the Last Word of Each String}
\usage{
findLastWord(string)
}
\arguments{
\item{string}{a character or character vector}
}
\value{
A character or character vector containing the last word of each entry
}
\description{
Find the Last Word of Each String
}
\examples{
findLastWord("hello world") # "world"
findLastWord(c("hello world", "hello mom")) # c("world", "mom")
}
