% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_harmonic.R
\name{fit_seasonal}
\alias{fit_seasonal}
\title{Fit yearly and bi-yearly sinusoids}
\usage{
fit_seasonal(t_months, prop)
}
\arguments{
\item{t_months}{a numeric vector indicating time (in months)}

\item{prop}{a numeric vector indicating the proportion of cases
in year reported during month}
}
\value{
a list containing seasonal coefficients and fitted values
}
\description{
Fit yearly and bi-yearly sinusoids
}
\examples{
t_months = 1:12
prop = c(0.1,0.1,0.2,0.2,0.3,0.3,0.3,0.2,0.3,0.2,0.2,0.1)
}
