% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{flatten_activities}
\alias{flatten_activities}
\title{turn a nested DF with one row per activity into something ready for plotting}
\usage{
flatten_activities(nested)
}
\description{
turn a nested DF with one row per activity into something ready for plotting
}
