% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatPvalue.R
\name{formatPvalue}
\alias{formatPvalue}
\title{formatPvalue}
\usage{
formatPvalue(x, digits = 2, n = 3, mode = "min_digits", Min = 1e-06)
}
\arguments{
\item{x}{a vector}

\item{digits}{reserve how many digits. default is 2}

\item{n}{The threshold of the scientific notion}

\item{mode}{When set to min, format as < Min; When set to min_digits, set to Min/10}

\item{Min}{the Min threshold to change fmt}
}
\value{
formated vector
}
\description{
format pvalue
}
\examples{
formatPvalue()

}
\author{
Rui Chen (crotoc@gmail.com)
}
\keyword{format}
\keyword{pvalue}
