% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAlignPlots.R
\name{g_legend}
\alias{g_legend}
\title{g_legend}
\usage{
g_legend(a.gplot)
}
\arguments{
\item{a.gplot}{}
}
\value{
A grob of legend
}
\description{
Return a legend from a ggplot
}
\examples{
g_legend()

}
\author{
Rui Chen (crotoc@gmail.com)
}
\keyword{Graphics}
\keyword{hplot}
