% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_oz.R
\name{geom_oz}
\alias{geom_oz}
\title{Geometry of Australia coast and state lines}
\usage{
geom_oz(color = "gray", fill = NULL, ...)
}
\arguments{
\item{color}{color of lines}

\item{fill}{fill color for mainland and Tasmania. Default is NULL (no fill)}

\item{...}{Other arguments to geom_path}
}
\value{
a geom object
}
\description{
Geometry of Australia coast and state lines
}
