% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_rect2.R
\name{geom_rect2}
\alias{geom_rect2}
\title{geom rect with alpha for tranparency only}
\usage{
geom_rect2(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", ..., na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)
}
\description{
geom rect with alpha for tranparency only
}
