% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{get_activity_data}
\alias{get_activity_data}
\title{Get activity data}
\usage{
get_activity_data(id)
}
\arguments{
\item{id}{Garmin Activity ID}
}
\value{
dataframe with one row per moment recorded
}
\description{
Get activity data
}
