% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_REMIND_run_names.R
\name{get_REMIND_run_names}
\alias{get_REMIND_run_names}
\title{get_REMIND_run_names}
\usage{
get_REMIND_run_names(paths)
}
\arguments{
\item{paths}{A vector of stings made up of either:
\itemize{
\item Names of the REMIND output-folders,
\item Paths to the REMIND output-folders,
\item Paths to files within the REMIND output-folders.
}}
}
\value{
A vector of strings with the names of the REMIND runs.
}
\description{
get_REMIND_run_names returns the names of REMIND runs.
}
