% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW3T1.R
\name{get_width_ci}
\alias{get_width_ci}
\title{Calculate the widths of the uncertainty intervals for a country}
\usage{
get_width_ci(est, iso_code, coverage)
}
\arguments{
\item{est}{A data frame containing the estimated values}

\item{iso_code}{The ISO code of the country}

\item{coverage}{The coverage level of the uncertainty interval (80 or 95)}
}
\value{
A data frame with the year and the corresponding interval width
}
\description{
Calculate the widths of the uncertainty intervals for a country
}
