% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPageWithProxy.R
\name{getPageWithProxy}
\alias{getPageWithProxy}
\title{getPageWithProxy}
\usage{
getPageWithProxy(url, proxy = "socks5h://127.0.0.1:1081")
}
\arguments{
\item{url}{}

\item{proxy}{}
}
\description{
getPageWithProxy
}
