% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTexStr.R
\name{getTexStr}
\alias{getTexStr}
\title{getTexStr}
\usage{
getTexStr(dt, width = 0, height = 0, splitN = 4, tablestr = "sidewaystable")
}
\arguments{
\item{method2order}{method to order colors (\code{"hsv"} or \code{"cluster"})}
}
\value{
None
}
\description{
Description: format tab to Tex string
}
\examples{
getTexStr

}
\author{
Rui Chen (crotoc@gmail.com)
}
