% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_encoding_cn.R
\name{guess_encoding_cn}
\alias{guess_encoding_cn}
\title{guess_encoding_cn
猜测中文文件的编码，并且返回概率最大的一个。}
\usage{
guess_encoding_cn(x)
}
\arguments{
\item{x}{文件路径}
}
\value{
编码
}
\description{
guess_encoding_cn
猜测中文文件的编码，并且返回概率最大的一个。
}
