% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_if_null_or_not_null.R
\name{ifnotnull}
\alias{ifnotnull}
\title{ifnotnull}
\usage{
ifnotnull(x, not.null.cmd, null.cmd)
}
\arguments{
\item{x}{A object to check}

\item{not.null.cmd}{A clojure{} to eval when x is not null}

\item{null.cmd}{A clojure{} to eval when x is null; can be miss.}
}
\description{
do someting if x is not null
}
