% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_if_null_or_not_null.R
\name{ifnull}
\alias{ifnull}
\title{ifnull}
\usage{
ifnull(x, cmd)
}
\arguments{
\item{x}{A object to check}

\item{cmd}{A clojure{} to eval when x is null}
}
\description{
do someting if x is null
}
