% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_my_local.R
\name{install_my_local}
\alias{install_my_local}
\title{install_my_local}
\usage{
install_my_local(path, pkg)
}
\arguments{
\item{path}{the path of the package}

\item{pkg}{the package name}
}
\value{
None
}
\description{
Description: Used to quickly install my pakages after modifications.
}
\examples{
install_my_local

}
\author{
Rui Chen (crotoc@gmail.com)
}
