% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installLib.R
\name{installLib}
\alias{installLib}
\title{installLib}
\usage{
installLib(type)
}
\arguments{
\item{method2order}{method to order colors (\code{"hsv"} or \code{"cluster"})}
}
\value{
None
}
\description{
Description: install all libs that used in my R script
}
\examples{
installLib(type=c("all"))
installLib(type=c("general","plot"))

}
\author{
Rui Chen (crotoc@gmail.com)
}
