% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isDateString.R
\name{isDateString}
\alias{isDateString}
\title{isDateString
Check if a string a Date format string.}
\usage{
isDateString(x)
}
\arguments{
\item{x}{A string should be a date like "2017-01-01".}
}
\value{
If x is a date stirng, True; else False.
}
\description{
isDateString
Check if a string a Date format string.
}
\examples{
isDateString("2017-01-01")
## [1] True
isDateString("20170101")
## [1] FALSE 
}
