% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{latlon_bb}
\alias{latlon_bb}
\title{grab the bounding box from a dataframe with lat/long columns}
\usage{
latlon_bb(df, lat = df$lat, lon = df$lon)
}
\description{
grab the bounding box from a dataframe with lat/long columns
}
