% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{lcm_}
\alias{lcm_}
\title{Least Common Multiple}
\usage{
lcm_(...)
}
\arguments{
\item{a}{a natural number}

\item{b}{a natural number}
}
\value{
least common multiple
}
\description{
Least Common Multiple
}
