% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_raw.r
\name{load_data_raw}
\alias{load_data_raw}
\title{Basic utilities for the project, reads the raw data
and splits the data into train and test data}
\usage{
load_data_raw()
}
\value{
list(train_data, test_data)
}
\description{
Basic utilities for the project, reads the raw data
and splits the data into train and test data
}
\examples{
load_data_raw()
}
