% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_vector.r
\name{load_data_vector}
\alias{load_data_vector}
\title{Basic utilities for the project, reads the raw data, converts them to one hot vector
and splits the data into train and test data}
\usage{
load_data_vector()
}
\value{
list(train_data, test_data)
}
\description{
Basic utilities for the project, reads the raw data, converts them to one hot vector
and splits the data into train and test data
}
\examples{
load_data_vector()
}
