% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myUtils.R
\name{mapIdsMouse}
\alias{mapIdsMouse}
\title{Map gene ids between one another for mouse}
\usage{
mapIdsMouse(IDs, IDFrom, IDTo)
}
\arguments{
\item{IDs}{The IDs to be mapped}

\item{IDFrom}{The current ID format, such as SYMBOL, ENSEMBL}

\item{IDTo}{The ID type to map to, such as SYMBOL, ENSEMBL}
}
\description{
This function maps ids such as symbols , entrez ids, ensembls ids between each other
}
