% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_morlet.R
\name{mat_2_df}
\alias{mat_2_df}
\title{small function to format the matrix into a dataframe}
\usage{
mat_2_df(mat_wt)
}
\arguments{
\item{mat_wt}{matrix containing the wavelet transform}
}
\value{
the dataframe in a long format
}
\description{
small function to format the matrix into a dataframe
}
