% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{math_gcd}
\alias{math_gcd}
\title{Calculate greatest common divisor by using Euclidean algorithm}
\usage{
math_gcd(...)
}
\arguments{
\item{a}{a natural number}

\item{b}{a natural number}
}
\value{
greatest common divisor
}
\description{
Calculate greatest common divisor by using Euclidean algorithm
}
