% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modis_fc_thredds_urls.R
\name{modis_fc_thredds_urls}
\alias{modis_fc_thredds_urls}
\title{create list of MODIS FC download links}
\usage{
modis_fc_thredds_urls(mod_h, mod_v, years)
}
\arguments{
\item{mod_h}{numeric vector of modis horizontal cells of interest}

\item{mod_v}{numeric vector of modis vertical cells of interest}

\item{years}{numeric vector of years of interest. i.e. 2001:2024, or c(2020, 2024) etc}
}
\value{
a character vector of download links
}
\description{
author: Ned Ryan-Schofield
date: 5/02/2026

Helper function for getting urls of MODIS fractional cover dataset from the thredds database
Catalogue url to browse the server is https://thredds.nci.org.au/thredds/catalog/catalog.html. To download, have to use https://thredds.nci.org.au/thredds/fileServer/... instead of catalog
Product of interest is Rangeland and Pasture Productivity RaPP = tc43
some tiles are not .006.nc, they are .061.nc. MODIS v6 discontinued feb2023, and have switched to v6.1. So use v6.1 when available (since 2020. Function automatically replaces with v6.1 for years >= 2020
}
