% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{month_first_day}
\alias{month_first_day}
\title{First day (day of year) of month}
\usage{
month_first_day()
}
\value{
a numeric vector of first day of month (day of year)
}
\description{
First day (day of year) of month
}
