% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{month_first_day_label}
\alias{month_first_day_label}
\title{First day of month}
\usage{
month_first_day_label()
}
\value{
a character vector of first day of month (day of year)
}
\description{
First day of month
}
