% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.help-package.R
\docType{package}
\name{move.help-package}
\alias{move.help}
\alias{move.help-package}
\title{move.help: movement analysis helper functions}
\description{
Various helper functions to import, export, and coerce movement data into formats required by different packages. Created March 2026, still needs some work.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nedschofield/move.help}
  \item Report bugs at \url{https://github.com/nedschofield/move.help/issues}
}

}
\author{
\strong{Maintainer}: Ned Ryan-Schofield \email{ned.ryan-schofield@cdu.edu.au}

}
\keyword{internal}
