% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{mse}
\alias{mse}
\title{Average squared difference}
\usage{
mse(x, y)
}
\arguments{
\item{x}{variable x}

\item{y}{variable y}
}
\value{
Average squared difference between two numeric vectors.
}
\description{
Average squared difference
}
\examples{
mse(runif(10), runif(10))
}
