% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{myApply}
\alias{myApply}
\title{Apply Functions Over Array Margins}
\usage{
myApply(x, MARGIN, FUN, ...)
}
\arguments{
\item{x}{an array, including a matrix.}

\item{MARGIN}{a vector giving the subscripts which the function will be applied over.
Or a character vector for dimension names will be applied over}

\item{FUN}{the function to be applied.}

\item{...}{optional arguments to FUN.}
}
\description{
Apply Functions Over Array Margins
}
