% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{myArrayGen}
\alias{myArrayGen}
\title{Generate a new array a vector to a array for a fixed dimension}
\usage{
myArrayGen(arr, vector, idx)
}
\arguments{
\item{arr}{arr}

\item{vector}{idx}
}
\description{
Generate a new array a vector to a array for a fixed dimension
}
