% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{myDownloadSILO}
\alias{myDownloadSILO}
\title{Download SILO weather records}
\usage{
myDownloadSILO(
  stations,
  output,
  start = "1889/1/1",
  end = format(Sys.time(), "\%Y/\%m/\%d")
)
}
\arguments{
\item{stations}{The number of weather station}

\item{output}{The output folder}

\item{start}{The start time of weather records}

\item{end}{The end time of weather records}
}
\description{
Download SILO weather records
}
