% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{myDoy2Date}
\alias{myDoy2Date}
\title{Convert doy to date}
\usage{
myDoy2Date(doy, format = "\%d-\%b")
}
\arguments{
\item{doy}{a vertor of day of year}

\item{format}{the date format}
}
\description{
Convert doy to date
}
