% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{myFindPpdStation}
\alias{myFindPpdStation}
\title{Get information of closest PPD weather station}
\usage{
myFindPpdStation(
  latitude,
  longitude,
  file = project_filepath("Results/weather/ncWbPpd.nc", project = "com")
)
}
\arguments{
\item{latitude}{latitude}

\item{file}{Path of netcdf file}

\item{longtitude}{longitude}
}
\description{
Get information of closest PPD weather station
}
