% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{myGetPpd}
\alias{myGetPpd}
\title{Get PPD weather records by station}
\usage{
myGetPpd(
  number = NULL,
  latitude = NULL,
  longitude = NULL,
  yrange = NULL,
  file = "../000-CommonFiles/Results/weather/ncWbPpd.nc"
)
}
\arguments{
\item{number}{The number of weather station}

\item{latitude}{latitude}

\item{yrange}{Year range}

\item{file}{Path of netcdf file}

\item{longtitude}{longitude}
}
\description{
Get PPD weather records by station
}
