% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{myList2Array}
\alias{myList2Array}
\title{Generate an empty array with a list}
\usage{
myList2Array(lst)
}
\arguments{
\item{lst}{A list to generate an array}
}
\description{
Generate an empty array with a list
}
