% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{myLmSlope}
\alias{myLmSlope}
\title{The the slope of linear regression}
\usage{
myLmSlope(y, x = NULL)
}
\arguments{
\item{y}{a vector of response.}

\item{x}{a vector of model variables. If null, fill along with y.}
}
\description{
The the slope of linear regression
}
