% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{myoption}
\alias{myoption}
\title{myoption}
\usage{
myoption()
}
\arguments{
\item{no}{param need.}
}
\value{
None
}
\description{
All options I have used.
}
\examples{
option <- myoption()
}
\author{
Rui Chen (crotoc@gmail.com)
}
