% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{myPlantUML}
\alias{myPlantUML}
\title{Convert plantuml into figure}
\usage{
myPlantUML(uml, text)
}
\arguments{
\item{uml}{PlantUML text}

\item{text}{Text to show}
}
\description{
Convert plantuml into figure
}
