% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication.R
\name{myPrintFigures}
\alias{myPrintFigures}
\title{Print figures into files including pdfs and pngs}
\usage{
myPrintFigures(
  p,
  filename = NULL,
  height = NULL,
  width = NULL,
  format = c("pdf", "png"),
  fontsize = 10,
  fun = NULL
)
}
\arguments{
\item{p}{the figure}

\item{filename}{the output filename without extension}

\item{height}{Figure height}

\item{width}{Figure width}

\item{format}{The output formats}

\item{fontsize}{The font size in the figures}

\item{fun}{The function to export figures}
}
\description{
Print figures into files including pdfs and pngs
}
