% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{myRunParallel}
\alias{myRunParallel}
\title{Title}
\usage{
myRunParallel(x, fun, cpus = parallel::detectCores() - 1, ...)
}
\arguments{
\item{x}{}

\item{fun}{}

\item{...}{}
}
\value{

}
\description{
Title
}
