% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mySubsetArr}
\alias{mySubsetArr}
\title{Subset array by value of dimensions}
\usage{
mySubsetArr(x, ...)
}
\arguments{
\item{x}{object to be subsetted.}

\item{...}{values of dimension}
}
\description{
Subset array by value of dimensions
}
