% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myutils.R
\docType{package}
\name{myutils}
\alias{myutils}
\alias{myutils-package}
\title{myutils}
\description{
A set of functions written to make data munging and preparation easier
}
\author{
Shyam Saladi, \email{saladi@caltech.edu}
}
