% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physics.R
\name{myutils_internal_constants}
\alias{myutils_internal_constants}
\title{Internal Constants in myutils}
\description{
Internal constants defined in 'myutils' package. Three colons (`myutils:::`) are necessary to access them.
}
\details{
Physical constants:
\itemize{
  \item g : Gravity Acceleration (= 9.8 m/s2)
  \item R : Gas Constant (J/K/mol)
  \item kB : Boltzmann constant (J/K)
  \item Pstp : Standard Temperature (= 100000 Pa)
  \item Tstp : Standard Pressure    (= 273.15 K)
  \item Psatp : Standard Ambient Temperature
  \item Tsatp : Standard Ambient Pressure
  \item Pntp : Normal Temperature
  \item Tntp : Normal Pressure
}
}
