% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_unique_cubic_spline.R
\name{N_4}
\alias{N_4}
\title{Definition of the unit cubic B-spline (centered at x_0, width sigma, amplitude A)}
\usage{
N_4(x, x_0, A, sigma)
}
\arguments{
\item{x}{a numeric vector where to evaluate the unit cubic spline}
}
\value{
evaluation of the unit cubic spline at values in \code{x}
}
\description{
Definition of the unit cubic B-spline (centered at x_0, width sigma, amplitude A)
}
\examples{
x <- seq(-5,5,by = 0.1)
plot(x,N_4_original(x),ylim = c(0,1))
lines(x,N_4_unit(x),col='red')
lines(x,N_4(x,x_0 = 1,A = 0.4,sigma = 2),col='green')
}
