% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_unique_cubic_spline.R
\name{N_4_original}
\alias{N_4_original}
\title{Definition of the cubic B-spline (cardinal B-spline of order 4)}
\usage{
N_4_original(x)
}
\arguments{
\item{x}{a numeric vector where to evaluate the cubic spline}
}
\value{
evaluation of the cubic spline at values in \code{x}
}
\description{
Definition of the cubic B-spline (cardinal B-spline of order 4)
}
\examples{
N_4_original(2)
}
