% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_unique_cubic_spline.R
\name{N_4_unit}
\alias{N_4_unit}
\title{Definition of the unit cubic B-spline (centered at 0, width 1, amplitude 1)}
\usage{
N_4_unit(x)
}
\arguments{
\item{x}{a numeric vector where to evaluate the unit cubic spline}
}
\value{
evaluation of the unit cubic spline at values in \code{x}
}
\description{
Definition of the unit cubic B-spline (centered at 0, width 1, amplitude 1)
}
\examples{
N_4(0)
}
