% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/myutils.r
\name{naValueGraph}
\alias{naValueGraph}
\title{naValueGraph function helps us to display NA values
map of a data frame.}
\usage{
naValueGraph(data, percent = 100)
}
\arguments{
\item{data}{is a data frame}

\item{percent}{is the percentage of row values to display
The percentage must be a number between 1 and 100}
}
\description{
naValueGraph function helps us to display NA values
map of a data frame.
}
\examples{
naValueGraph(data=df, percent=50)
}
\keyword{NA}
\keyword{data}
\keyword{frames}
\keyword{values,}

