% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nci_thredds_download.R
\name{nci_thredds_download}
\alias{nci_thredds_download}
\title{download files from NCI thredds database}
\usage{
nci_thredds_download(
  download_urls,
  out_dir,
  overwrite = FALSE,
  timeout = 180,
  quiet = FALSE
)
}
\arguments{
\item{download_urls}{character vector of download links starting with format above}

\item{out_dir}{character string of output file directory location. Function creates it if it doesn't exist}
}
\value{
downloads each file individually, and storing them in the output directory provided.
}
\description{
author: Ned Ryan-Schofield
date: 5/02/2026
Function for downloading files from the NCI thredds database (or other data repositories probably). Catalogue url to browse the server is https://thredds.nci.org.au/thredds/catalog/catalog.html
To download, have to use https://thredds.nci.org.au/thredds/fileServer/... instead of catalog
}
