% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeOutput.R
\name{normalizeOutput}
\alias{normalizeOutput}
\title{normalize output directory}
\usage{
normalizeOutput(opt)
}
\arguments{
\item{method2order}{method to order colors (\code{"hsv"} or \code{"cluster"})}
}
\value{
None
}
\description{
Description: normalize directory
}
\examples{
normalize output directory

}
\references{

}
\seealso{

}
\author{
Rui Chen (crotoc@gmail.com)
}
\keyword{}
