% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{nrmse}
\alias{nrmse}
\title{Normalized root mean square error (nrmse)}
\usage{
nrmse(x, y, method = c("rangeobs", "sdobs", "meanobs", "interquartileobs"))
}
\arguments{
\item{x}{variable x}

\item{y}{variable y}

\item{method}{method to calculate nrmse. Valid values are:
\itemize{
    \item{sdobs: standard deviation of observations (x)}
    \item{rangeobs: difference between maximum and minimum observations (x)}
    \item{meanobs: average of observations (x)}
    \item{interquartileobs: the difference between 25th and 75th percentile (x)}
}}
}
\value{
Normalized root mean square error.
}
\description{
Normalized root mean square error (nrmse)
}
\examples{
nrmse(runif(10), runif(10))
}
