% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alucard.R
\name{numeric_summary}
\alias{numeric_summary}
\title{A summary of a numeric vector}
\usage{
numeric_summary(x, na.rm = T, verbose = F)
}
\arguments{
\item{x}{a numeric vector}

\item{na.rm}{an optional logical parameter. \code{TURE} by default}

\item{verbose}{is TRUE for verbose output}
}
\value{
A named vector with six values
}
\description{
Custom version of \code{summary} function for a numeric vector
}
\examples{
\dontrun{
x <- c(4,7,8)
numeric_summary(x)
}
}
