% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{nwordsfun}
\alias{nwordsfun}
\title{Determine the number of words in a given string}
\usage{
nwordsfun(str)
}
\arguments{
\item{str}{a character}
}
\value{
A numeric indicating how many words are in the input.
}
\description{
Determine the number of words in a given string
}
\examples{
nwordsfun("hello world") # 2
}
