% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opendev.R
\name{opendev}
\alias{opendev}
\title{opendev}
\usage{
opendev(opt)
}
\arguments{
\item{opt}{opt list from optparse
opt$ext pdf, jpeg, any device
opt$width
opt$height
opt$res resolution
opt$output output name}
}
\value{
None
}
\description{
Creates a plot of the crayon colors in \code{\link{brocolors}}
}
\examples{
opendev()

}
\author{
Rui Chen (crotoc@gmail.com)
}
\keyword{plot}
