% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myutils.R
\name{p_errorBar}
\alias{p_errorBar}
\title{Add a error bar}
\usage{
p_errorBar(x0, y0, x1, y1, col = "black", length = 0.1)
}
\arguments{
\item{x0, y0}{coordinates of points from which to draw}

\item{col}{color of error bar}

\item{length}{length of the edges of the error bar (in inches)}

\item{x0, y1}{coordinates of points to which to draw. At least one must the supplied}
}
\description{
Add a error bar
}
