% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_base.R
\name{p_polygon_error}
\alias{p_polygon_error}
\title{Highlight a region within a specific error}
\usage{
p_polygon_error(x, y, err_percentage, col = "#00000028", lty = 0)
}
\arguments{
\item{x, y}{vector of x or y}

\item{err_percentage}{percentage of error}

\item{lty}{type of the border line [optional]}
}
\description{
Highlight a region within a specific error
}
