% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_base.R
\name{p_polygon_upper_lower}
\alias{p_polygon_upper_lower}
\title{Highlight a region defined by upper and lower limits of profile data}
\usage{
p_polygon_upper_lower(x, y_upper, y_lower, col = "#00000028", lty = 0)
}
\arguments{
\item{x}{x (vector)}

\item{y_upper, y_lower}{upper and lower limits of y (vector)}

\item{col}{color of the highlighted region [optional]}

\item{lty}{type of the border line [optional]}
}
\description{
Highlight a region defined by upper and lower limits of profile data
}
