% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{pairwise}
\alias{pairwise}
\title{Apply a function to pairwise of all combinations in row or column of a matrix.}
\usage{
pairwise(X, margin = c(1, 2), FUN, ...)
}
\arguments{
\item{X}{a matrix}

\item{margin}{row (1) or column (2) to apply the function}

\item{FUN}{the function to be applied, which has two arguments
and only returns a single value.}

\item{...}{optional arguments to FUN}
}
\value{
a matrix with the same number of row (margin = 1) or column (margin = 2)
with the original matrix.
}
\description{
Apply a function to pairwise of all combinations in row or column of a matrix.
}
\examples{
X <- matrix(c(1, 2, 3, 4,
              1, 2, 5, 6,
              1, 2, 5, 7), nrow = 4)
FUN <- function(x, y, na.rm = FALSE) {
   sum(x != y, na.rm = na.rm)
}
pairwise(X, 2, FUN, na.rm = TRUE)
}
