% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{pairwise_diff}
\alias{pairwise_diff}
\title{Calculate the pairwise difference of all combinations in row or column of a matrix.}
\usage{
pairwise_diff(X, margin = c(1, 2), na.rm = FALSE)
}
\arguments{
\item{X}{a matrix}

\item{margin}{row (1) or column (2) to apply the function}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
a matrix with the same number of row (margin = 1) or column (margin = 2)
with the original matrix.
}
\description{
Calculate the pairwise difference of all combinations in row or column of a matrix.
}
\examples{
X <- matrix(c(1, 2, 3, 4,
              1, 2, 5, 6,
              1, 2, 5, 7), nrow = 4)
pairwise_diff(X, 2)
}
