% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{paste_gg}
\alias{paste_gg}
\title{... = ggplot objects}
\usage{
paste_gg(
  ...,
  fout,
  save_heights = 5,
  save_widths = 5,
  row_heights = NULL,
  row_widths = NULL,
  total_height = 5,
  total_width = 10,
  nrow,
  ncol
)
}
\description{
... = ggplot objects
}
