% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{paste_png}
\alias{paste_png}
\title{... = png file paths}
\usage{
paste_png(
  ...,
  fout,
  width,
  height,
  nrow,
  ncol,
  widths = NULL,
  heights = NULL,
  units = "pixels",
  just = "centre"
)
}
\description{
... = png file paths
}
