% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_morlet.R
\name{phase_angle}
\alias{phase_angle}
\title{extract phase from mat_wt}
\usage{
phase_angle(mat_wt)
}
\arguments{
\item{mat_wt}{matrix containing the wavelet transform}
}
\value{
the time series of phase angles
}
\description{
extract phase from mat_wt
}
