% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{pkg_info}
\alias{pkg_info}
\title{Get information for packages on CRAN.}
\usage{
pkg_info(pkgname, baseurl = NULL)
}
\arguments{
\item{pkgname}{Package Name}

\item{baseurl}{Baseurl for package info, defaults to "https://cran.r-project.org/web/packages/"}
}
\value{
Package information. (NULL if CRAN doesn't have it.)
}
\description{
Get information for packages on CRAN.
}
\examples{
pkg_info("tidyverse") #=> Returns information.
pkg_infO("R") #=> Returns NULL

}
