% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_CEStree.R
\name{plot_ces_parameters}
\alias{plot_ces_parameters}
\title{plot_ces_parameters}
\usage{
plot_ces_parameters(data, in_all_in, in_cesParameter, run_names = NULL)
}
\arguments{
\item{data}{A tibble with the pm_cesData for REMIND runs. E.g. the result of
\code{\link[=read_items_from_gdxs]{read_items_from_gdxs()}}.}

\item{in_all_in}{A string with the all_in parameter to be plotted.}

\item{in_cesParameter}{A string with the cesParameter to be plotted.}

\item{run_names}{A string of REMIND run names, or NULL.}
}
\value{
A ggplot2 object.
}
\description{
plot_ces_parameters does this and that
}
