% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cp.R
\name{plot_cp}
\alias{plot_cp}
\title{This function creates a visualization of cp in a dataset.}
\usage{
plot_cp(est, observed_data, iso_code, CI = NULL)
}
\arguments{
\item{data}{A data frame or matrix containing the data to analyze.}

\item{change_points}{A numeric vector indicating the positions of CP.}

\item{title}{A character string for the plot title. Defaults to "CP Plot".}

\item{color}{A character string specifying the color of change point lines. Defaults to "red".}
}
\value{
A ggplot object displaying the data.
}
\description{
This function creates a visualization of cp in a dataset.
}
