% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_morlet.R
\name{plot_re}
\alias{plot_re}
\title{plot real part of the wavelet decomposition}
\usage{
plot_re(mat_wt)
}
\arguments{
\item{mat_wt}{matrix containing the wavelet transform}
}
\value{
ggplot
}
\description{
plot real part of the wavelet decomposition
}
