% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_CEStree.R
\name{plot_sankey}
\alias{plot_sankey}
\title{Title}
\usage{
plot_sankey(cesOut2cesIn, cesIO, my_run, my_reg, my_period, style)
}
\arguments{
\item{cesOut2cesIn}{t}

\item{cesIO}{t}

\item{my_run}{t}

\item{my_reg}{t}

\item{my_period}{t}

\item{style}{t}
}
\value{
A sankeyNetwork object
}
\description{
Title
}
