% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPLabelFormat.R
\name{plotPLabelFormat}
\alias{plotPLabelFormat}
\title{plotPLabelFormat}
\usage{
plotPLabelFormat(n, threshold = 2.2e-16, digits = 2)
}
\arguments{
\item{n}{pvalue vector}

\item{threshold}{label "< threshold" if lower than threshold}

\item{digits}{reserve n digits}
}
\value{
a vector for labeling p value
}
\description{
Create beautiful multiple pages plot from ggplot object.
}
\examples{
out <- plotPLabelFormat(pvalue)

}
\author{
Rui Chen (crotoc@gmail.com)
}
\keyword{plot}
