% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyPaginate.R
\name{prettyPaginate}
\alias{prettyPaginate}
\title{prettyPaginate}
\usage{
prettyPaginate(myplot, opt)
}
\arguments{
\item{myplot}{a ggplot object}

\item{opt}{options from optparse
opt$type boxploth/pointh or others
opt$rowfacet
opt$facet_cname
opt$ncol
opt$nrow
opt$scale
opt$space
opt$x.variable
opt$facet_grid  formular}
}
\value{
a grob list
}
\description{
Create beautiful multiple pages plot from ggplot object.
}
\examples{
out <- prettyPaginate(x_plot,opt)

}
\author{
Rui Chen (crotoc@gmail.com)
}
\keyword{plot}
