% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myutils.R
\name{read.matrix}
\alias{read.matrix}
\title{Read matrix data from a csv file.}
\usage{
read.matrix(csvfile, skip = 0)
}
\arguments{
\item{csvfile}{path of a csvfile}

\item{skip}{number of lines}
}
\description{
Read matrix data from a csv file.
}
