% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_items_from_mifs.R
\name{read_items_from_mifs}
\alias{read_items_from_mifs}
\title{read_items_from_mifs}
\usage{
read_items_from_mifs(
  mif_filepaths,
  regexs_of_items,
  returnAsMagpie = FALSE,
  n_cores = 1
)
}
\arguments{
\item{mif_filepaths}{A vector of strings containing the paths to the mif
files with the raw data.}

\item{regexs_of_items}{A vector of regex patterns identifying the variables
to extract from the mif files.}

\item{returnAsMagpie}{Logical indicating return type. Either list of quitte
objects, or magpie.}

\item{n_cores}{Integer indicating the number of cores to be used in parallel}
}
\value{
my_data Either a list of quitte objects (default) or a single magpie
object,
}
\description{
read_items_from_mifs reads specific items from the remind reporting mif
files.
}
