% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{rmse}
\alias{bias}
\alias{mse}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{Metrics}{\code{\link[Metrics]{bias}}, \code{\link[Metrics]{mse}}, \code{\link[Metrics]{rmse}}}
}}

