% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refClassMap.R
\name{refClassMap}
\alias{refClassMap}
\title{A 2D map from 2D input to any output class. Note that keys must be numerics, and values are typechecked.}
\description{
A 2D map from 2D input to any output class. Note that keys must be numerics, and values are typechecked.
}
\section{Fields}{

\describe{
\item{\code{.class}}{The class of the output}
}}

