% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{regFormula}
\alias{regFormula}
\title{Build a regression formula for the dependent variable by the covariates, 
parsing form text form}
\usage{
regFormula(depVar, covariates)
}
\arguments{
\item{depVar}{a string, the dependent variable}

\item{covariates}{a string or vector of strings, the covariates}
}
\value{
The regression formula
}
\description{
Build a regression formula for the dependent variable by the covariates, 
parsing form text form
}
