% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{replaceCurlies}
\alias{replaceCurlies}
\title{Fix annoying encoding issues}
\usage{
replaceCurlies(txt)
}
\arguments{
\item{txt}{a character}
}
\value{
The same character but curly quotes and apostrophes subbed for their straight equivalents
}
\description{
Fix annoying encoding issues
}
