% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceNAwithZero.R
\name{replaceZerowithNA}
\alias{replaceZerowithNA}
\title{replaceNAwithZero}
\usage{
replaceZerowithNA(DT)
}
\arguments{
\item{method2order}{method to order colors (\code{"hsv"} or \code{"cluster"})}
}
\value{
None
}
\description{
Description:
}
\examples{
replaceNAwithZero

}
\author{
Rui Chen (crotoc@gmail.com)
}
