% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myUtils.R
\name{rpkm}
\alias{rpkm}
\title{Counts per million normalization}
\usage{
rpkm(data, geneLengths)
}
\arguments{
\item{data}{The count matrix to be normalized}

\item{geneLengths}{The gene lengths in the samer order as the rows of the count matrix}
}
\description{
This function converts an unnormalized countmatrix of genes x samples into read-per-kilobase normalization
}
\examples{
rpkm()
}
\keyword{normalization}
