% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{rrmse}
\alias{rrmse}
\title{Relative root mean square error}
\usage{
rrmse(x, y)
}
\arguments{
\item{x}{variable x}

\item{y}{variable y}
}
\value{
Relative root mean square error.
}
\description{
Relative root mean square error
}
\examples{
rrmse(runif(10), runif(10))
}
