% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{rsq}
\alias{rsq}
\title{Squared coefficient of correlation (R2)}
\usage{
rsq(x, y)
}
\arguments{
\item{x}{variable x}

\item{y}{variable y}
}
\value{
Squared coefficient of correlation (R2) or coefficient of determination.
}
\description{
Squared coefficient of correlation (R2)
}
\examples{
rsq(runif(10), runif(10))
}
