% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{run_parallel}
\alias{run_parallel}
\title{A wrap function to run parallel in different platforms}
\usage{
run_parallel(x, fun, cpus = parallel::detectCores() - 1, ...)
}
\arguments{
\item{x}{A numeric vector for parallel}

\item{fun}{A function}

\item{cpus}{Number of CPUs for snowfall package}

\item{...}{Other arguments passed to fun}
}
\value{
Result of parallel function.
}
\description{
A wrap function to run parallel in different platforms
}
\details{
The supported platforms include
* local computer through snowfall package (default)
* MPI cluster through snow package
* Windows HPC pack for parameter sweep job
}
\examples{
\dontrun{
fun <- function(i, a) {
  return(i * a)
}
run_parallel(seq(1, 10), fun, a = 2)
}
}
