% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{setEnvVarsFromFile}
\alias{setEnvVarsFromFile}
\title{Sets environment variables from a file}
\usage{
setEnvVarsFromFile(file)
}
\arguments{
\item{file}{The path to the file containing the variables to add. File must be a list of variables formatted "varname=varvalue" with no quotes, one per line}
}
\description{
Sets environment variables from a file
}
