% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_n_levels}
\alias{show_n_levels}
\title{show_n_levels}
\usage{
show_n_levels(df)
}
\arguments{
\item{df}{input dataframe}
}
\value{
list
}
\description{
calculate number of levels in each categorical field
}
\examples{
data(iris)
show_n_levels(iris)
}
