% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_quantile}
\alias{show_quantile}
\title{show_quantile}
\usage{
show_quantile(df, q = seq(0, 1, 0.1))
}
\arguments{
\item{df}{dataframe that quantile function performed on}

\item{q}{quantiles, defaults to 0 to 1, with step of 0.1}
}
\value{
list
}
\description{
calculate the quantiles for all numeric columns
}
\examples{
data(iris)
show_quantile(iris)
}
