% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{slope}
\alias{slope}
\title{Slope of linear regression}
\usage{
slope(x, y)
}
\arguments{
\item{x}{a numeric vector for x}

\item{y}{a numeric vector for y}
}
\value{
Slope of linear regression
}
\description{
Slope of linear regression
}
\examples{
slope(runif(10), runif(10))
}
