% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_dir.R
\name{sourceDir}
\alias{sourceDir}
\title{sourceDir}
\usage{
sourceDir(path, trace = TRUE, recursive = T, ...)
}
\arguments{
\item{path}{path to the source directory}

\item{trace}{print the path if True}

\item{recursive}{}

\item{...}{}
}
\description{
source all .R files in a directory
}
