% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_utils.R
\name{sp2df}
\alias{sp2df}
\title{Transforms a sp object into a dataframe to plot into ggplot}
\usage{
sp2df(x_sp)
}
\arguments{
\item{x_sp}{the sp object to transform into dataframe}
}
\value{
the corresponding dataframe to be plotted using geom_polygon
}
\description{
Transforms a sp object into a dataframe to plot into ggplot
}
