% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_unique_cubic_spline.R
\name{spline_optim}
\alias{spline_optim}
\title{Function optimizing the timing, spread and amplitude of peak}
\usage{
spline_optim(t, y)
}
\arguments{
\item{t}{numeric vector of time values}

\item{y}{numeric vector of observed values at time \code{t}}
}
\value{
the residual sum of squares between N_4(t,x_0,A,sigma) and y
}
\description{
Function optimizing the timing, spread and amplitude of peak
}
