% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ss_bins}
\alias{ss_bins}
\title{2006 Author Hideaki Shimazaki
Department of Physics, Kyoto University
shimazaki at ton.scphys.kyoto-u.ac.jp
http://toyoizumilab.brain.riken.jp/hideaki/res/histogram.html}
\usage{
ss_bins(x)
}
\description{
2006 Author Hideaki Shimazaki
Department of Physics, Kyoto University
shimazaki at ton.scphys.kyoto-u.ac.jp
http://toyoizumilab.brain.riken.jp/hideaki/res/histogram.html
}

