% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states.R
\name{stateNameToAbb}
\alias{stateNameToAbb}
\title{Translate between states (plus DC) and their postal abbreviations}
\usage{
stateNameToAbb(states)
}
\arguments{
\item{states}{a character}
}
\value{
States translated to postal abbreviations, or reverse
}
\description{
Translate between states (plus DC) and their postal abbreviations
}
