% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_begins_and_ends.R
\name{strBeginsWith}
\alias{strBeginsWith}
\title{strBeginsWith}
\usage{
strBeginsWith(x, y)
}
\arguments{
\item{x}{A big string should begins with y}

\item{y}{A small string}
}
\value{
If x begins with y, then TRUE; else then FALSE
}
\description{
String x begins with y
}
\examples{
strBeginsWith("hello","he")
## [1] TRUE
strBeginsWith("hello","lo")
## [1] FALSE
}
