% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_begins_and_ends.R
\name{strEndsWith}
\alias{strEndsWith}
\title{strEndsWith}
\usage{
strEndsWith(x, y)
}
\arguments{
\item{x}{A big string should ends with y}

\item{y}{A small string}
}
\value{
If x ends with y, then TRUE; else then FALSE
}
\description{
String x ends with y
}
\examples{
strEndsWith("hello","lo")
## [1] TRUE
strEndsWith("hello","he")
## [1] FALSE
}
