% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{summarise_model}
\alias{summarise_model}
\title{Summarise a model with statistics indicators}
\usage{
summarise_model(
  data,
  x = "x",
  y = "y",
  digits = NULL,
  direction = c("wide", "long"),
  extra = FALSE,
  .groups = "drop",
  ...
)
}
\arguments{
\item{data}{A data frame to summarise}

\item{x}{x variable name. Supports bare column names (e.g. \code{x}) and strings (e.g. \code{"x"}).}

\item{y}{y variable name. Supports bare column names (e.g. \code{y}) and strings (e.g. \code{"y"}).}

\item{digits}{integer indicating the number of decimal places (round) or significant digits (signif) to be used.}

\item{direction}{the wide (default) or long format for the output}

\item{extra}{summarise extra variables. \code{FALSE} in default.}

\item{.groups}{Grouping structure passing to \code{summarise} function. "drop" in default.}

\item{...}{other arguments passing to functions. Supporting arguments
\itemize{
    \item{nrmse_method: Method for nrmse}
}}
}
\value{
A data frame with statistics indicators in columns including:
\itemize{
    \item{n: Number of rows}
    \item{r: Coefficient of correlation}
    \item{r2: Squared coefficient of correlation}
    \item{bias: Average amount by which actual is greater than predicted}
    \item{mse: Average squared difference}
    \item{rmse: Root mean squared error}
    \item{nrmse: Normalized root mean squared error}
    \item{d: Willmott degree of agreement}
    \item{error7day: Percentabe of errors less than 7 days if \code{extra} is \code{TRUE}}
}
}
\description{
Summarise a model with statistics indicators
}
\examples{
library(dplyr)
data <- data.frame(x = 1:10, y = 1:10 + runif(10))
data |> summarise_model()
data |> summarise_model(x = x, y = y)
data |> summarise_model(x = "x", y = "y")
data |> summarise_model(digits = 2)
# Export as long format
data |> summarise_model(digits = 2, direction = "long")
}
