% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabsets.R
\name{tabsets}
\alias{tabsets}
\title{Create tabsets for R Markdown and Quarto documents}
\usage{
tabsets(data, FUN, ..., use_quarto = TRUE)
}
\arguments{
\item{data}{A data frame grouped by one or more variables using `dplyr::group_by()`.}

\item{FUN}{A function that generates the content for each tab.}

\item{...}{Arguments passed to `FUN`.}

\item{use_quarto}{Logical; if `TRUE`, generates Quarto-compatible tabsets}
}
\value{
A character vector or HTML structure representing the tabsets
}
\description{
This function generates tabsets that can be used in R Markdown or Quarto 
documents to organize content into tabs. It provides a convenient way to 
create multiple tabs with headers and content.
}
\details{
This function should be used within R Markdown chunks with 
`results="asis"` to properly render the tabset markup.
}
\examples{
iris |> 
  dplyr::group_by(Species) |>
  tabsets(FUN = knitr::kable)
iris |> 
  dplyr::group_by(Species) |>
  tabsets(FUN = knitr::kable, use_quarto = FALSE)
}
