% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{trim_GBcode}
\alias{trim_GBcode}
\title{remove '_cnt' or '_pref' from GBcodes}
\usage{
trim_GBcode(vect_names)
}
\arguments{
\item{vect_names}{a vector with names to clean}
}
\value{
vact_names with 'GBcode' instead of 'GBcode_cnt' or 'GBcode_pref'
}
\description{
remove '_cnt' or '_pref' from GBcodes
}
