% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_morlet.R
\name{ts_power}
\alias{ts_power}
\title{plot real part of the wavelet decomposition}
\usage{
ts_power(mat_wt, dates_seq, range_scale)
}
\arguments{
\item{mat_wt}{matrix containing the wavelet transform}
}
\value{
ggplot
}
\description{
plot real part of the wavelet decomposition
}
