% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/using.R
\name{using}
\alias{using}
\title{using}
\usage{
using(...)
}
\arguments{
\item{packages}{names}
}
\value{
None
}
\description{
Install package if not installed. From BiocManager.
}
\examples{
using("ggplot2")

}
\author{
Rui Chen (crotoc@gmail.com)
}
