% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myutils.R
\name{v_closestValue}
\alias{v_closestValue}
\title{Get the closest value(s) to `target`}
\usage{
v_closestValue(v, target)
}
\arguments{
\item{v}{a vector}

\item{target}{a number}
}
\value{
closest value(s) to `target`
}
\description{
Get the closest value(s) to `target`
}
