% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myutils.R
\name{v_cum_probability}
\alias{v_cum_probability}
\title{Create a matrix of information of cumulative distribution.}
\usage{
v_cum_probability(values, decreasing = FALSE)
}
\arguments{
\item{values}{vector}

\item{decreasing}{If TRUE, the values are ordered from largest to smallest (optional)}
}
\value{
matrix
}
\description{
If you want to create a graph of cumulative distribution, you should use
ecdf() function.
  > fCP <- ecdf( c(3,76,58,24,100,1) ); plot(fCP)
}
