% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view}
\alias{view}
\title{view}
\usage{
view(table, n = 5)
}
\arguments{
\item{table}{data.frame, matrix, data.table ...}

\item{n}{how many columns to display}
}
\value{
None
}
\description{
Creates a plot of the crayon colors in \code{\link{brocolors}}
}
\examples{
view(dt)

}
\author{
Rui Chen (crotoc@gmail.com)
}
\keyword{plot}
