% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewNoName.R
\name{viewNoName}
\alias{viewNoName}
\title{viewNoName}
\usage{
viewNoName(table, n = 5)
}
\arguments{
\item{table}{data.frame, matrix, data.table ...}

\item{n}{how many columns to display}
}
\value{
None
}
\description{
view table without row and colnames
}
\examples{
viewNoName(dt)

}
\author{
Rui Chen (crotoc@gmail.com)
}
\keyword{view}
