% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whos.R
\name{whos}
\alias{whos}
\title{Whos: show objects' info in current env}
\usage{
whos(no_functions = T)
}
\arguments{
\item{no_functions}{}
}
\value{
object names
}
\description{
Whos: show objects' info in current env
}
\examples{
whos()
}
