% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{write_dataset}
\alias{write_dataset}
\title{Used for model training}
\usage{
write_dataset(df, filename, outcome_col = "outcome")
}
\description{
groups should be specified within the dataframe provided
using dplyr (i.e., group_by(df))
}
