% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_morlet.R
\name{wt_extract}
\alias{wt_extract}
\title{Extract components between 2 scales}
\usage{
wt_extract(mat_wt, range_scale = NULL, range_date = NULL, fill = 0)
}
\arguments{
\item{mat_wt}{matrix containing the wavelet transform}

\item{range_scale}{vector of length 2 specifying the range of scales}

\item{range_date}{vector of 2 dates specifying the range of dates}

\item{fill}{variable to fill matrix outside of scale range}
}
\value{
wt matrix with scales in range_scale and zero otherwise
}
\description{
Extract components between 2 scales
}
