% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_morlet.R
\name{wt_inverse}
\alias{wt_inverse}
\title{Reconstruct from wavelet transform}
\usage{
wt_inverse(mat_wt, y.m = 0, J = 110)
}
\arguments{
\item{mat_wt}{matrix containing the wavelet transform}

\item{y.m}{(optional) mean of original time series}

\item{J}{determines the largest scale s_J = s_0 2^(J dj)}
}
\value{
the inverse of the wavelet decomposition
}
\description{
Reconstruct from wavelet transform
}
