% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{xlsx2csv}
\alias{xlsx2csv}
\title{Convert xlsx file into csv.}
\usage{
xlsx2csv(xlsx, path)
}
\arguments{
\item{xlsx}{An Excel file.}

\item{path}{Destination path.}
}
\value{
Contents of the input.
}
\description{
Convert xlsx file into csv.
}
