% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{bisection}
\alias{bisection}
\title{Bisection Method}
\usage{
bisection(f, rangeFrom, rangeTo, itMax = 100, tol = 1e-07)
}
\arguments{
\item{f}{function returning a real value}

\item{rangeFrom}{maximum value of range}
}
\value{
root
}
\description{
Bisection Method
}
