% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_morlet.R
\name{calc_power}
\alias{calc_power}
\title{calculate wavelet power spectrum from mat_wt}
\usage{
calc_power(mat_wt)
}
\arguments{
\item{mat_wt}{matrix containing the wavelet transform}
}
\value{
the matrix of wavelet power spectrum
}
\description{
calculate wavelet power spectrum from mat_wt
}
