% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_summaries.R
\name{char_summary}
\alias{char_summary}
\title{Summaries for character vectors}
\usage{
char_summary(x)
}
\arguments{
\item{x}{vector.}
}
\value{
a numeric vector of length, number of missing values and number of
unique values.
}
\description{
Summaries for character vectors
}
\examples{
cvec <- c("Boston", NA, "Brookline", "Brighton", NA, "Boston")
char_summary(cvec)
}
