% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_utilities.R
\name{char_summary}
\alias{char_summary}
\title{A summary of a character vector}
\usage{
char_summary(x, na.rm = TRUE)
}
\arguments{
\item{x}{a character vector}

\item{na.rm}{an optional logical parameter. \code{TRUE} by default.}
}
\value{
A named vector with three values.
}
\description{
Custom version of \code{summary} function for a character vector.
}
\examples{
\dontrun{
x <- c("A", "B", "A", "C")
char_summary(x)
}
}
