% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_summaries.R
\name{char_summary}
\alias{char_summary}
\title{Summary for a character vector}
\usage{
char_summary(x, na.rm = FALSE)
}
\arguments{
\item{x}{character vector}

\item{na.rm}{whether to ignore NA values}
}
\value{
a numeric vector with 3 values
}
\description{
Display summary for a character value
}
\examples{

x <- c("Boston", "New York", "Boston", "New Haven")
char_summary(x)

}
