% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fip.R
\name{DarcyWeisbach}
\alias{DarcyWeisbach}
\title{Darcy-Weisbach equation}
\usage{
DarcyWeisbach(fD, density, velocity, D)
}
\arguments{
\item{fD}{Darcy friction factor}

\item{density}{density (kg/m3)}

\item{velocity}{velocity (m/s)}

\item{D}{pipe diameter (m)}
}
\value{
dP per unit length (Pa/m)
}
\description{
Darcy-Weisbach equation
}
