% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_to_movebank_tab.R
\name{DT_to_movebank_tab}
\alias{DT_to_movebank_tab}
\title{data.table to movebank textfile}
\usage{
DT_to_movebank_tab(
  dt,
  out.path,
  t = "timestamp",
  x = "long",
  y = "lat",
  ID = "ID"
)
}
\arguments{
\item{dt}{a list of data.tables that are gps data}

\item{x}{numerical variable - data.table coordinate longitude column}

\item{y}{numerical variable - data.table coordinate latitude column}

\item{ID}{character variable - data.table animal ID column}

\item{outpath}{character variable of file path name. i.e. "./outputs/table/tiwi_cats_movebank_up_03032026.txt"}

\item{timestamp}{POSIXct variable - data.table timestamp column}
}
\value{
writes a textfile to disc.
}
\description{
author: Ned Ryan-Schofield
date: 03/03/2026
Helper function for coercing a list of data.table objects of gps data that have been checked for errors and coerced back to data.table to a single .txt file for upload to movebank.
}
