% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_CEStree.R
\name{explore_CEStree}
\alias{explore_CEStree}
\title{Explore the CES tree of REMIND}
\usage{
explore_CEStree(gdx_filepaths)
}
\arguments{
\item{gdx_filepaths}{A vector of strings to REMIND gdx files}
}
\description{
Opens a shiny app in which the CES parameters and values may be explored.
}
