% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fip.R
\name{fD_Blasius}
\alias{fD_Blasius}
\title{Blasius correlation to calculate Darcy friction factor}
\usage{
fD_Blasius(Re, C = 0.3164)
}
\arguments{
\item{Re}{Reynolds number}

\item{C}{constant}
}
\value{
Darcy friction factor
}
\description{
Blasius correlation to calculate Darcy friction factor
}
