% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fip.R
\name{fD_Colebrook}
\alias{fD_Colebrook}
\title{Colebrook correlation to calculate Darcy friction factor}
\usage{
fD_Colebrook(roughness, D, Re, tol = 1e-08, itMax = 10, warn = TRUE)
}
\description{
Colebrook correlation to calculate Darcy friction factor
}
