% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_oz.R
\name{geom_oz}
\alias{geom_oz}
\title{Geometry of Australia coast and state lines}
\usage{
geom_oz(color = "gray", ...)
}
\arguments{
\item{color}{color of lines}

\item{...}{Other arguments to geom_path}
}
\value{
a geom object
}
\description{
Geometry of Australia coast and state lines
}
