% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_oz.R
\name{geom_oz_states}
\alias{geom_oz_states}
\title{Geom label for Australian States}
\usage{
geom_oz_states(color = "gray", ACT = FALSE)
}
\arguments{
\item{color}{color for text}

\item{ACT}{Whether to include ACT (FALSE in default)}
}
\value{
geom_text
}
\description{
Geom label for Australian States
}
