% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{math_lcm}
\alias{math_lcm}
\title{Least Common Multiple}
\usage{
math_lcm(...)
}
\arguments{
\item{a}{a natural number}

\item{b}{a natural number}
}
\value{
least common multiple
}
\description{
Least Common Multiple
}
