% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{model_summarise}
\alias{model_summarise}
\title{Summarise a model with statistics indicators}
\usage{
model_summarise(
  data,
  x = "x",
  y = "y",
  digits = NULL,
  direction = c("wide", "long")
)
}
\arguments{
\item{data}{A data frame to summarise}

\item{x}{x variable name}

\item{y}{y variable name}

\item{digits}{integer indicating the number of decimal places (round) or significant digits (signif) to be used.}

\item{direction}{The wide (default) or long format for the output}
}
\value{
A data frame with statistics indicators in columns including:
\itemize{
    \item{n: Number of rows}
    \item{r: Coefficient of correlation}
    \item{r2: Squared coefficient of correlation}
    \item{bias: Average amount by which actual is greater than predicted}
    \item{mse: Average squared difference}
    \item{rmse: Root mean squared error}
    \item{rrmse: Relative root mean squared error}
}
}
\description{
Summarise a model with statistics indicators
}
\examples{
library(dplyr)
data <- data.frame(x = 1:10, y = 1:10 + runif(10))
data \%>\% model_summarise()
data \%>\% model_summarise(digits = 2)
# Export as long format
data \%>\% model_summarise(digits = 2, direction = "long")
}
