% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{model_summarise}
\alias{model_summarise}
\title{Summarise a model with statistics indicators}
\usage{
model_summarise(data, x = "x", y = "y")
}
\arguments{
\item{data}{A data frame to summarise}

\item{x}{x variable name}

\item{y}{y variable name}
}
\value{
A data frame with statistics indicators in columns
}
\description{
Summarise a model with statistics indicators
}
\examples{
library(dplyr)
data <- data.frame(x = 1:10, y = 1:10)
data \%>\% model_summarise()
}
