% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myUtils-package.R
\docType{package}
\name{myUtils-package}
\alias{myUtils-package}
\alias{myUtils}
\title{Load data from multiple gdx and mif files.}
\description{
Load data from multiple gdx and mif files. Analyse and compare REMIND results.
}
