% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myUtils-package.R
\docType{package}
\name{myUtils-package}
\alias{myUtils}
\alias{myUtils-package}
\title{myUtils: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
