% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{newtonRaphson}
\alias{newtonRaphson}
\title{Newton-Raphson method}
\usage{
newtonRaphson(fun, dFun, fD_0, tol = 1e-10, itMax = 10)
}
\arguments{
\item{fun}{function returning a real value}

\item{dFun}{derivative of \code{fun}}

\item{fD_0}{initial value of the argument of \code{fun}}

\item{tol}{desired accuracy (convergence tolerance)}

\item{itMax}{the maximum number of iterations}
}
\value{
root of \code{fun}
}
\description{
Newton-Raphson method
}
