% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm-LAPTOP-FP3FHU3F.R, R/algorithm.R
\name{newtonRaphson}
\alias{newtonRaphson}
\title{Newton-Raphson method}
\usage{
newtonRaphson(fun, dFun, x_0, tol = 1e-10, itMax = 10)

newtonRaphson(fun, dFun, x_0, tol = 1e-10, itMax = 10)
}
\arguments{
\item{fun}{function returning a real value}

\item{dFun}{derivative of \code{fun}}

\item{x_0}{initial value of the argument of \code{fun}}

\item{tol}{desired accuracy (convergence tolerance)}

\item{itMax}{the maximum number of iterations}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ root of `fun`
}\if{html}{\out{</div>}}

root of \code{fun}
}
\description{
Newton-Raphson method

Newton-Raphson method
}
\seealso{
\itemize{
\item \code{\link[=bisection]{bisection()}}
}

\itemize{
\item \code{\link[=bisection]{bisection()}}
}
}
