% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_summaries.R
\name{numeric_summary}
\alias{numeric_summary}
\title{Summaries for numeric vectors}
\usage{
numeric_summary(x, na.rm = FALSE)
}
\arguments{
\item{x}{vector.}

\item{na.rm}{passed into other functions.}
}
\value{
a numeric vector with min, max, mean, sd, length and number of
missing values.
}
\description{
Summaries for numeric vectors
}
\examples{
nvec <- c(2022, 2021, NA, 2021, 2021, NA, 2022)
numeric_summary(nvec, na.rm = TRUE)
}
