% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_utilities.R
\name{numeric_summary}
\alias{numeric_summary}
\title{A summary of a numeric vector}
\usage{
numeric_summary(x, na.rm = TRUE)
}
\arguments{
\item{x}{a numeric vector}

\item{na.rm}{an optional logical parameter. TRUE by default.}
}
\value{
A named vector with six values.
}
\description{
Custom version of summary function for a numeric vector.
}
\examples{
x <- c(4,7,8)
numeric_summary(x)
}
