% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myutils.R
\name{p_blank}
\alias{p_blank}
\title{Create a blank plot}
\usage{
p_blank(
  xRange,
  yRange,
  axes = TRUE,
  log = "",
  xlab = "",
  ylab = "",
  main = "",
  xaxs = "r",
  yaxs = "r",
  las = 0,
  cex.axis = 1,
  cex.lab = 1
)
}
\arguments{
\item{xRange, yRange}{Ranges of plot}

\item{axes}{Draw axes when TRUE}

\item{log}{a character string ("", "x", "y", or "xy")}

\item{xlab, ylab}{Label name of x or y axis}

\item{main}{Overall title}

\item{xaxs, yaxs}{Extra space}

\item{las}{the style of axis labels (default = 1)}

\item{cex.axis}{Font size of axis annotation (relative to the current setting of cex)}

\item{cex.lab}{Font size of labels}
}
\description{
Create a blank plot
}
