% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_base.R
\name{p_blank_excel}
\alias{p_blank_excel}
\title{Create a Excel-like blank plot}
\usage{
p_blank_excel(
  xRange,
  yRange,
  xlab = "",
  ylab = "",
  cex.axis = 1.25,
  cex.lab = 1.25
)
}
\arguments{
\item{xRange, yRange}{ranges of plot}

\item{xlab, ylab}{label names}

\item{cex.axis, cex.lab}{cexs of axis and lab}
}
\description{
Create a Excel-like blank plot
}
