% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myutils.R
\name{p_blank_fit}
\alias{p_blank_fit}
\title{Blank plot with axes that fits within the original data range}
\usage{
p_blank_fit(
  xRange,
  yRange,
  axes = TRUE,
  xlab = "",
  ylab = "",
  main = "",
  las = 1,
  cex.axis = 1.25
)
}
\arguments{
\item{xRange, yRange}{ranges of plot}

\item{axes}{Draw axes when TRUE}

\item{xlab, ylab}{Label name of x or y axis}

\item{main}{Overall title}
}
\description{
Blank plot with axes that fits within the original data range
}
