% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_base.R
\name{p_blank_fit}
\alias{p_blank_fit}
\title{Blank plot with axes that fits within the original data range}
\usage{
p_blank_fit(
  xRange,
  yRange,
  axes = TRUE,
  log = "",
  xlab = "",
  ylab = "",
  main = "",
  las = 1,
  cex.axis = 1.25,
  cex.lab = 1
)
}
\arguments{
\item{xRange, yRange}{Ranges of plot}

\item{axes}{Draw axes when TRUE}

\item{log}{a character string ("", "x", "y", or "xy")}

\item{xlab, ylab}{Label name of x or y axis}

\item{main}{Overall title}

\item{las}{the style of axis labels (default = 1)}

\item{cex.axis}{Font size of axis annotation (relative to the current setting of cex)}

\item{cex.lab}{Font size of labels}
}
\description{
Blank plot with axes that fits within the original data range
}
