% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_base.R
\name{p_col32}
\alias{p_col32}
\title{Add alpha level to color}
\usage{
p_col32(cols, alphas)
}
\arguments{
\item{cols}{vector of color name or hex (e.g. red, #123456)}

\item{alphas}{vector of alpha levels (from 0 to 255) [optional]}
}
\value{
vector of color codes
}
\description{
Add alpha level to color
}
\examples{
cc <- p_col32(c("#123456", "red"), c(96,128))
cat(cc)   # #12345660 #FF000080

}
