% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyRUtilities.R, R/myutils.R
\name{p_errorBarX}
\alias{p_errorBarX}
\title{Add a error bar along X axis}
\usage{
p_errorBarX(x, y, err, col = "black", length = 0.1)

p_errorBarX(x, y, err, col = "black", length = 0.1)
}
\arguments{
\item{x, y}{coordinates of the center point of the error bar}

\item{err}{value of error}

\item{col}{color of error bar}

\item{length}{length of the edges of the error bar (in inches)}
}
\description{
Add a error bar along X axis

Add a error bar along X axis
}
\seealso{
[p_errorBar()]

[p_errorBar()]
}
