% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myutils.R
\name{p_errorBarY}
\alias{p_errorBarY}
\title{Add a error bar along Y axis}
\usage{
p_errorBarY(x, y, err, col = "black", length = 0.1)
}
\arguments{
\item{x, y}{coordinates of the center point of the error bar}

\item{err}{value of error}

\item{col}{color of error bar}

\item{length}{length of the edges of the error bar (in inches)}
}
\description{
Add a error bar along Y axis
}
\seealso{
[p_errorBar()]
}
