% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_base.R
\name{p_value2col}
\alias{p_value2col}
\title{Convert a value to a hex color with a color pallet}
\usage{
p_value2col(v, vmin, vmax, f_pal = heat.colors, n = 256, log10_scale = FALSE)
}
\arguments{
\item{v}{vector of value(s)}

\item{vmin, vmax}{minimum and maximum of `v`}

\item{f_pal}{color pallet (function)}

\item{n}{number of steps}

\item{log10_scale}{xxx}
}
\description{
Convert a value to a hex color with a color pallet
}
