% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myutils.R
\name{p_value2col}
\alias{p_value2col}
\title{convert a value to a hex color with a color pallet}
\usage{
p_value2col(v, vmin, vmax, f_pal = heat.colors, n = 256, log10_scale = FALSE)
}
\description{
convert a value to a hex color with a color pallet
}
