% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cp.R
\name{plot_cp}
\alias{plot_cp}
\title{This function returns a ggplot with or without confidence interval.}
\usage{
plot_cp(est, observed_data, iso_code, CI = NULL)
}
\arguments{
\item{est}{ibble which contains mCPR estimates. Columns: “Country or area”, iso, Year, Median, U95,L95}

\item{iso_code}{the countrys iso code number}

\item{data}{A olumns: iso, year, cp}

\item{COnfidence}{Interval can be NA or number}
}
\value{
A ggplot object displaying the data.
}
\description{
This function returns a ggplot with or without confidence interval.
}
