% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myutils.R
\name{rcbmat}
\alias{rcbmat}
\title{Read clipboard as a data frame by using \code{utils::read.table()}}
\usage{
rcbmat(header, ...)
}
\arguments{
\item{header}{If TRUE, retrieved data contains a header}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{  A data frame containing clipboard data
}\if{html}{\out{</div>}}
}
\description{
Read clipboard as a data frame by using \code{utils::read.table()}
}
