% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyRUtilities.R
\name{rcbmat}
\alias{rcbmat}
\title{Read clipboard as a data frame by using \code{utils::read.table()}}
\usage{
rcbmat(header, ...)
}
\arguments{
\item{header}{If TRUE, retrieved data contains a header}
}
\value{
A data frame containing clipboard data
}
\description{
Read clipboard as a data frame by using \code{utils::read.table()}
}
