% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_items_from_gdxs.R
\name{read_items_from_gdxs}
\alias{read_items_from_gdxs}
\title{read_items_from_gdxs}
\usage{
read_items_from_gdxs(gdx_filepaths, gdx_items, remind_names = TRUE)
}
\arguments{
\item{gdx_filepaths}{A vector of strings with the paths to all gdx files.}

\item{gdx_items}{A vector of strings}

\item{remind_names}{If true, assume filepaths to remind output folders/fulldata.gdx files.}
}
\value{
A list of tibbles with the data for each gdx_item requested.
}
\description{
Extract select items from gdxs.
}
\examples{
\dontrun{
read_items_from_gdxs(gdx_filepaths = c("file1.gdx", "file2.gdx"),
                     gdx_items = c("item1", "item2.l", "item3.m"))}
}
