% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_items_from_mifs.R
\name{read_items_from_mifs}
\alias{read_items_from_mifs}
\title{read_items_from_mifs}
\usage{
read_items_from_mifs(mif_filepaths, regexs_of_items, returnAsMagpie = FALSE)
}
\arguments{
\item{mif_filepaths}{A vector of strings containing the paths to the mif
files with the raw data.}

\item{regexs_of_items}{A vector of regex patterns identifying the variables
to extract from the mif files.}

\item{returnAsMagpie}{If TRUE return magclass object}
}
\value{
my_data Either a list of tibbles (default) or a single magpie object.
}
\description{
read_items_from_mifs reads specific items from the remind reporting mif
files.
}
