% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myutils.R
\name{wcbmat}
\alias{wcbmat}
\title{Copy table data to clipboard with \code{utils::write.table()}}
\usage{
wcbmat(
  data,
  header,
  sep = "\\t",
  size = 128,
  row.names = FALSE,
  qmethod = "double",
  col.names = ifelse(header && row.names, NA, header),
  ...
)
}
\arguments{
\item{data}{data.frame or matrix}

\item{header}{xx}

\item{sep}{the field separator string}

\item{size}{clipboard size}

\item{row.names}{yy}

\item{qmethod}{description}

\item{col.names}{description}

\item{...}{optionals}
}
\description{
Copy table data to clipboard with \code{utils::write.table()}
}
